Refactor feature management code:
- Split `FeatureFlagsEndpointFilter` implementation and extension methods into separate files for better modularity.
- Updated feature flag validation to use `IVariantFeatureManagerSnapshot` with cancellation token support.
- Adjusted the filter's return logic to use `Results.NotFound()` when the feature is disabled.
- Simplified the `WithFeatureGate` extension method by removing the generic `TBuilder` type.
- Improved code organization and enhanced documentation clarity.
