+++ src/Microsoft.FeatureManagement.AspNetCore/FeatureFlagsEndpointFilterExtensions.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 
 namespace Microsoft.FeatureManagement.AspNetCore
 {
-
     /// <summary>
     /// Extension methods that provide feature management integration for ASP.NET Core endpoint building.
     /// </summary>
@@ -17,7 +16,6 @@ namespace Microsoft.FeatureManagement.AspNetCore
         /// </summary>
         /// <param name="builder">The endpoint convention builder.</param>
         /// <param name="featureName">The name of the feature flag to evaluate.</param>
-        /// <typeparam name="TBuilder">The type of the endpoint convention builder.</typeparam>
         /// <returns>The endpoint convention builder for chaining.</returns>
         /// <remarks>
         /// This extension method enables feature flag control over endpoint access. When the feature is disabled,
@@ -30,53 +28,9 @@ namespace Microsoft.FeatureManagement.AspNetCore
         ///     .WithFeatureGate("MyFeature");
         /// </code>
         /// </example>
-        public static TBuilder WithFeatureGate<TBuilder>(this TBuilder builder, string featureName)
-            where TBuilder : IEndpointConventionBuilder
+        public static IEndpointConventionBuilder WithFeatureGate(this IEndpointConventionBuilder builder, string featureName)
         {
             return builder.AddEndpointFilter(new FeatureFlagsEndpointFilter(featureName));
-        }
-    }
-
-    /// <summary>
-    /// An endpoint filter that requires a feature flag to be enabled.
-    /// </summary>
-    public class FeatureFlagsEndpointFilter : IEndpointFilter
-    {
-        public string FeatureName { get; }
-
-        /// <summary>
-        /// Creates a new instance of <see cref="FeatureFlagsEndpointFilter"/>.
-        /// </summary>
-        /// <param name="featureName">The name of the feature flag to evaluate for this endpoint.</param>
-        public FeatureFlagsEndpointFilter(string featureName)
-        {
-            if (string.IsNullOrEmpty(featureName))
-            {
-                throw new ArgumentNullException(nameof(featureName));
-            }
-
-            FeatureName = featureName;
-        }
-
-        /// <summary>
-        /// Invokes the feature flag filter to control endpoint access based on feature state.
-        /// </summary>
-        /// <param name="context">The endpoint filter invocation context containing the current HTTP context.</param>
-        /// <param name="next">The delegate representing the next filter in the pipeline.</param>
-        /// <returns>
-        /// A <see cref="Result.NotFound"/> if the feature is disabled, otherwise continues the pipeline by calling the next delegate.
-        /// Returns a ValueTask containing the result object.
-        /// </returns>
-        public async ValueTask<object> InvokeAsync(EndpointFilterInvocationContext context, EndpointFilterDelegate next)
-        {
-            IFeatureManagerSnapshot fm = context.HttpContext.RequestServices.GetRequiredService<IFeatureManagerSnapshot>();
-            if (fm is null)
-            {
-                return await next(context);
-            }
-
-            bool enabled = await fm.IsEnabledAsync(FeatureName);
-            return enabled ? await next(context) : Results.NotFound();
         }
     }
 }
